Clean up and document unused i18n token finder script
